looping over different values of P(r=0) and outputting total mass M(r=R).
as suggested in the GSL manual, I set the ODE function
return value to GSL_EBADFUNC if either 1.) the pressure at a given
radius is negative, or 2.) if either derivative yields a NaN (which
usually happens at the same radius where P < 0). this return value
causes the integrator to stop. I then dump the last result for
M(r), which should be M(r=R). However that seems not always to
work. In particular, if you printf M(r) and r in the ODE function
itself, as well as printf from the main program after the ODE
function has returned GSL_EBADFUNC, you'll find they don't match;
in particular, the integrator seems to progress several steps
farther in r than indicated by the "last value" printf. no idea
why this is.

however, even with this weird "feature," a plot of P(r = 0) and
M(r = R) looks good: there's definitely a global maximum M(R) which
hopefully will be the Chandra mass when I put the units back. however
there are a few huge spikes and dips which probably correspond to
that weird thing I just talked about.